requirements.txt
flask
folium
joblib
numpy
pandas
pillow
scikit-image
scikit-learn
torch
torchvision
pypdf
matplotlib
